+++ src/Umbraco.Community.Forms.Workflows/Client/src/contexts/mailerlite.config.context.token.ts
@@ -1,0 +1,4 @@
+import type { MailerLiteConfigContext } from "./mailerlite.config.context.js";
+import { UmbContextToken } from "@umbraco-cms/backoffice/context-api";
+
+export const MAILERLITE_CONFIG_CONTEXT = new UmbContextToken<MailerLiteConfigContext>("mailerlite-config-context");